ImpactChron

Documentation for ImpactChron.

ImpactChron.AsteroidHistory

ImpactChron.Nrm

ImpactChron.PetroTypes

ImpactChron.Unf

ImpactChron.lNrm

ImpactChron.agerecal

ImpactChron.csv2dict

ImpactChron.csv2nt

ImpactChron.data2csv

ImpactChron.dict2csv

ImpactChron.downscale!

ImpactChron.histogramify

ImpactChron.histogramify!

ImpactChron.impact_reset_array!

ImpactChron.js_arx2dict

ImpactChron.ll_dist

ImpactChron.ll_dist_params

ImpactChron.ll_param

ImpactChron.ll_params

ImpactChron.mcmean

ImpactChron.nan_regolith!

ImpactChron.perturb

ImpactChron.planetesimal_cooling_dates

ImpactChron.planetesimal_cooling_dates!

ImpactChron.planetesimal_cooling_timestep!

ImpactChron.planetesimal_temperature

ImpactChron.radius_at_depth

ImpactChron.rangemidbounds

ImpactChron.rangemidpoints

ImpactChron.tturbosum

ImpactChron.weight_petro_types!

ImpactChron.AsteroidHistory — Type

AsteroidHistory(typeseed<:Number; nnodes, Δt, tmax, downscale_factor)

struct containing Arrays that record the evolution of a (bombarded) asteroid.

Constructor function takes a parameter from the proposal to seed type and requires the number of radial nodes nnodes (::Int), the timestep used Δt, the full time duration tmax, and the downscale_factor (::Int).

Fields: Vfrxn: volume fractions of each radial shell, peakT: peak temperature of each radial shell, cooltime: indices in t of the primary cooling date, impacts: number of impacts at each timestep, txr: time x radius array of proportional cooling ages, agedist: distribution of ages corresponding to t, agedist_downscaled: distribution of ages corresponding to t_downscaled, t: timesteps of full-scale model, t_downscaled: timesteps of downscaled model output,

ImpactChron.Nrm — Type

Nrm(μ::Float64,σ::Float64)

Immutable struct to describe normally distributed data, reported as mean (μ) and 1σ (σ)

ImpactChron.PetroTypes — Type

PetroTypes(temps::NamedTuple,samples::Vector{String})

struct containing fields of type3-type6, reflecting petrologic types, each with subfilds T (maximum temperature in K) and p (proportion among the chondrite record). Note that ps sum to unity.

Constructor takes a NamedTuple containing maximum temperatures as fields T3-T6 and a Vector{String} containing the petrologic types corresponding to chondrites used as a prior.

If no argument given – PetroTypes() – returns zeroed type_ fields and weight=false, which short-circuits weighting by petrologic type.

ImpactChron.Unf — Type

Unf(a::Float64,b::Float64)

Immutable struct to describe uniformly distributed data, reported as minimum (a) and maximum (b).

ImpactChron.lNrm — Type

lNrm(μ::Float64,σ::Float64)

Immutable struct to describe lognormally distributed data, reported as log-space mean (μ) and 1σ (σ)

ImpactChron.agerecal — Method

agerecal(x,sig;monitor_age,n, KAr=false)

Recalculate the age and uncertainty of an Ar-Ar age x±sig (1σ) Ma with the decay constants of Steiger & Jager (1977). Quantifies uncertainty by resampling n times (default: 1M). Optionally accepts monitor_age. If not given, resamples from a range of absolute (not necessarily K-Ar constrained) monitor ages ∈ [0,3] Ga.

If KAr=true, this recalculates the K-Ar age with the new decay constants, without correcting for a monitor.

ImpactChron.csv2dict — Method

csv2dict(filename::String;symbol::Bool=true)

Read a .csv file into a NamedTuple. Assumes columns reflect individual entries, with the first row specifying that entry's key. For NaN-buffered columns (see nt2csv), the NaNs are removed, returning a single-element entry.

A true value for symbol will convert data saved as arrays of Strings into arrays of Symbols.

see also: nt2csv, csv2dict

ImpactChron.csv2nt — Method

csv2nt(filename::String;symbol::Bool=true)

Read a .csv file into a NamedTuple. Assumes columns reflect individual elements, with the first row specifying that element's key. For NaN-buffered columns (see nt2csv), the NaNs are removed, returning a single-element entry.

A true value for symbol will convert data saved as arrays of Strings into arrays of Symbols.

see also: nt2csv, csv2dict

ImpactChron.data2csv — Method

data2csv(filename::String,data)

Save data to a .csv file. data may be a composite type of Dict or NamedTuple or an Array. Accepts fields of any type element. For single-element entries, the rest of the array/table is filled with NaNs.

see also: nt2csv, dict2csv

ImpactChron.dict2csv — Method

dict2csv(filename::String,D::Dict)

Save a Dict to a .csv. Accounts for fields of any type. For single-element entries, the rest of the array/table is filled with NaNs.

see also: nt2csv, data2csv

ImpactChron.downscale! — Method

downscale!(B::AbstractArray, A::AbstractArray)

Downscales elements of 1-D array A into smaller 1-D array B by summing. Scales the downscaled values in B by the ratio of length(A)÷length(B) to preserve any normalizations.

Requires that length(A) % length(B) = 0.

ImpactChron.histogramify! — Method

histogramify!(dist::AbstractVector, domain::AbstractRange, x::AbstractVector, y::AbstractVector)

In-place histogramify that overwites a pre-allocated vector dist.

see histogramify for details

ImpactChron.histogramify — Method

histogramify(domain::AbstractRange,x::AbstractVector,y::AbstractVector)

Constructs histogram over (linear) midpoints of domain from model outputs in x with corresponding abundances in y. Does not require a constant step in x.

Normalizes the output, such that for output dist ∑ dist[dᵢ] * Δd = 1 (for each dᵢ in the bincenters of domain with step-size Δd), so long as all x ∈ domain. If any x ∉ domain, ∑ dist[dᵢ] * Δd = 1- (∑yₒᵤₜ / ∑yₐₗₗ ) where the corresponding xₒᵤₜ of each yₒᵤₜ is ∉ domain.

Returns only the histogram masses, centers of time bins must be calculated externally. e.g.

Δd = step(domain)
bincenters= LinRange( first(domain)+Δd/2, last(domain)-Δd/2, length(domain)-1)

or see rangemidpoints

ImpactChron.impact_reset_array! — Method

function impact_reset_array!(tₓr::AbstractArray,solartime::AbstractArray,tcoolₒ::AbstractArray,Vfrxn::AbstractArray,
                                impacts::AbstractArray,
                                p::NamedTuple,c::NamedTuple;
                                nᵣ::Integer,Δt::Number)

Simulates an impact history from -χ parameters in p (below), and resets primary planetesimal cooling dates (indices of dates in solartime in tcoolₒ) and fractional volumes (Vfraxn) based on impact/crater properties described in c.

Depth-cooling age (relative) abundances are tracked in array tₓr (time x radial depth), with dimensions (time,radius) = (length(solartime),nᵣ), where nᵣ describes the number of radial nodes, as in the planetesimal_cooling_dates function.

impacts and tcoolₒ are pre-allocated vectors that respectively record the number of impacts at each time step and the index of the primary cooling date in solartime.

Impact site geometries are codified by types (Cone,Parabola,Hemisphere) defined in c and calculated in the radius_at_depth.

Impact flux follows an exponential decay described by parameters in p:

p.tχ ~ instability start time

p.τχ ~ e-folding timescale of impact flux

p.Fχ ~ initial impact flux

ImpactChron.js_arx2dict — Method

js_arx2dict(file::String, vars::Tuple; n, ll=true, accept=true, perturbation=false)

Load a serialized archive file from thermochron_metropolis into a Dictfor "post-run" analysis, only incorporating the variables invarsand coveringnsteps (all steps by default). Provide aBoolto incorporatell,accept, orperturbation`

ImpactChron.ll_dist — Method

ll_dist(x::AbstractVector,dist::AbstractVector,mu::AbstractVector,sigma::AbstractVector)

Calculate loglikelihood that observations in mu and sigma are drawn from modeled distribution described by x and dist where x contains the bincenters of a normalized histogram dist and mu and sigma respectively contain the mean and 1σ of the observations.

ImpactChron.ll_dist_params — Method

function ll_dist_params(a::AsteroidHistory, p, plims, mu,sig)

Calculate the combined log-likelihood that

Observations with mean mu and corresponding 1σ uncertainties sig (both ::Array) are drawn from the downscaled age distribution and timesteps contained in a (::AsteroidHistory)

The proposal parameters p used to calculate this history are drawn from the prior distributions plims (both ::NamedTuple)

If the age distribution is all zero, quickly returns a log-liklihood of -∞

ImpactChron.ll_param — Method

```julia ll_param(x::Number,D::T) -> T ∈ {Nrm,lNrm,Unf}

Calculate the log-likelihood that x is drawn from a distribution D, where D may be... Normal (Nrm), with mean D.μ and 1σ = D.σ

Lognormal (lNrm), with logspace mean D.μ and 1σ = D.σ. Assumes x is already in logspace.

Uniform (Unf) with lowerbound D.a and upperbound D.b (D::Unf always returns loglikelihhood of zero, bounds test is done earlier to speed up code.)

see ImCh_parameters.jl for construction of T-structs

ImpactChron.ll_params — Method

ll_params(p::NamedTuple,d::NamedTuple)

Calculate log-likelihood for a number of proposals in p with corresponding distributions in d

Currently does not evaluate impact (χ) parameters.

ImpactChron.mcmean — Method

mcmean(x,xsig,n=10_000_000,fullpost=false)

Calculate mean of age(s) x with 1σ uncertainties xsig by Monte Carlo method. A 10M resample (default) returns consistent results at the 10ka level.

Returns a tuple of (mean,1σ) by default.

If fullpost=true, returns posterior samples rather than summary statistics.

ImpactChron.nan_regolith! — Method

nan_regolith!(d::AbstractArray,T::AbstractArray,Tmin::Number)

Replace all dates in d with NaN where corresponding peak temperatures in T are less than Tmin Fast with @turbo

ImpactChron.perturb — Method

perturb(p::NamedTuple,k::Symbol,n::Number)

Return a NamedTuple identical to p, with one field (key k) changed to the value of n. Note that == identity is preserved only if the order of fields in p is as below

Fields: tss,rAlo,R,ta,cAl,Tm,Tc,ρ,Cp,k,tχα,τχα,Fχα,tχβ,τχβ,Fχβ

ImpactChron.planetesimal_cooling_dates! — Method

function planetesimal_cooling_dates!(ages::AbstractArray, Vfrxn::AbstractArray,peakT::AbstractArray, p::NamedTuple;
    nᵣ::Integer, Δt::Number, tmax::Number, Tmax::Number, Tmin::Number)

In-place planetesimal_cooling_dates that updates ages, Vfrxn, and peakT.

see also planetesimal_cooling_dates

ImpactChron.planetesimal_cooling_dates — Method

planetesimal_cooling_dates(p::NamedTuple; nᵣ, Δt, tmax, Tmax, Tmin)

Returns an NTuple{4,Vector} containing (in this order) the thermochronologic cooling dates (Ma after CAIs) and corresponding volume fractions, radial depths (km from center), and peak temperatures (K) of nᵣ evenly spaced nodes in a spherical body.

Physical and environmental parameters are described in p. Alternatively, these parameters may be individually listed in lieu of ϕ. These parameters are outlined in the table below. Note that several of these parameters need to be entered as the natural logarithm of the value for easy compatibility with the inversion function.

Δt gives the timestep (in Ma), tmaxdescribes the duration of the model (Ma after CAIs), andTmaxandTmindefine the maximum and minimum temperatures (K) allowed for chondritic material in the body. Default values are only given fortmax(2000 Ma),Tmax(1500 K), andTmin` (0 K).

| Parameter                 | log?  | `NmTpl` | `func`  |
| ------------------------- | ----  | ------ | -------- |
| closure temperature (K)   | no    | `Tc`   | `Tc`     |
| solar system age (Ma)     | no    | `tss`  | `tₛₛ`    |
| initial ²⁶Al/²⁷Al         | no    | `rAlo` | `rAlo`   |
| body radius (m)           | yes   | `R`    | `R`      |
| accretion date (Ma)       | yes   | `ta`   | `tₐ`     |
| disk temperature (K)      | yes   | `Tm`   | `To`     |
| [Al] (g/g)                | yes   | `cAl`  | `Al_conc`|
| density (kg/m³)           | yes   | `k`    | `K`      |
| thermal diffusivity       | yes   | `ρ`    | `ρ`      |
| spec. heat capacity       | yes   | `Cp`   | `Cₚ`     |
| ------------------------- | ----  | ------ | -------- |

see also: planetesimal_cooling_dates!

ImpactChron.planetesimal_cooling_timestep! — Method

function planetesimal_cooling_timestep!(solartime::AbstractRange, time_i::Vector, Vfrxn::Vector, peakT::Vector, p::NamedTuple; nᵣ, Tmax, Tmin)

Returns thermochronologic cooling dates in time_i as indices of solartime, along with corresponding volumetric fractions (Vfrxn) and peak temperatures in K (peakT) for nᵣ nodes in a body with planetesimal and environmental parameters given in p. Tmax and Tmin respectively describe the maximum and minimum temperatures allowed in the chondritic planetesimal. Failing to exceed Tmin gives the date of accretion, and exceeding Tmax sets the volumetric fraction to zero (achondritic).

see also: planetesimal_cooling_dates, planetesimal_cooling_dates!

ImpactChron.planetesimal_temperature — Method

function planetesimal_temperature(time::AbstractArray, radii::AbstractArray; To::Float64, Ao::Float64, λ::Float64, K::Float64, κ::Float64 )

Calculates the evolution of temperature at a range of depths defined for a conductively cooling sphere with thermal conductivity K and thermal diffusivity κ, given ambient temperature To, initial heat production Ao, and heat-producing-element decay constant 'λ'.

Adapted from: Carlslaw & Jäger (1959) and Hevey & Sanders (2006)

ImpactChron.radius_at_depth — Method

radius_at_depth(rᵢ::Number, R::Number, x::T) where T<:{Cone,Parabola,Hemisphere}

Calculates the radius of a circular area at a radial distance of rᵢ from the center of a body with radius R, where the volume of the region is approximated by a cone (x::Cone), paraboloid of rotation (x::Parabola), or hemisphere (x::Hemisphere). x includes a maximum depth (x.z) and surface radius (x.r). When x::Hemisphere, only r is used.

ImpactChron.rangemidbounds — Method

rangebinbounds(x::AbstractRange)

Calculate a LinRange of the linear bounds for each "midpoint" step in x.

ImpactChron.rangemidpoints — Method

rangemidpoints(x::AbstractRange)

Calculate a LinRange of the midpoints of each step in x(<:AbstractRange).

ImpactChron.tturbosum — Method

tturbosum(x::AbstractArray)

Fast summing of x with the power of LoopVectorization.jl's @tturbo (multithreaded turbo).

ImpactChron.weight_petro_types! — Method

weight_petro_types!(v::AbstractArray,T::AbstractArray,petrotypes::NamedTuple)

Reweight volumetric fractions relative to the abundance of each petrologic type in the meteorite record. Takes Vectors of volumetric fraction (v), peak temperature (T), and cooling date (d), as output by planetesimal_cooling_dates. Accounts for melted layers (r where v[i]==0 due to melting in planetesimal_cooling_timestep!. Requires all petrologic types to occupy at least one radial node, otherwise returns zero in all v.

petrotypes is a NamedTuple of NamedTuples, such that petrotypes = (type#=(T<:Number,p<:Number), ...) where T and p respectively represent maximum temperature and relative abundance in the meteorite record.