ImpactChron

Documentation for ImpactChron.

ImpactChron.Nrm

ImpactChron.Unf

ImpactChron.lNrm

ImpactChron.ImpactResetQuench

ImpactChron.area_at_depth

ImpactChron.csv2dict

ImpactChron.csv2nt

ImpactChron.data2csv

ImpactChron.dict2csv

ImpactChron.histogramify

ImpactChron.histogramify!

ImpactChron.impact_reset_array!

ImpactChron.ll_dist

ImpactChron.perturb

ImpactChron.rangemidpoints

ImpactChron.tturbosum

ImpactChron.Nrm — Type

Nrm(μ::Float64,σ::Float64)

Immutable struct to describe normally distributed data, reported as mean (μ) and 1σ (σ)

ImpactChron.Unf — Type

Unf(a::Float64,b::Float64)

Immutable struct to describe uniformly distributed data, reported as minimum (a) and maximum (b).

ImpactChron.lNrm — Type

lNrm(μ::Float64,σ::Float64)

Immutable struct to describe lognormally distributed data, reported as log-space mean (μ) and 1σ (σ)

ImpactChron.ImpactResetQuench — Method

This is the old ImpactRestAr. Needs to be largely rewritten to a framework similar to ImpactResetArray. Most importantly, replace probabilistic maths with number of hits at each timestep maths.

CURRENTLY NOT FUNCTIONAL

FORMER DOCUMENTATION:

ImpactResetAr ~ reheat volumes for an exponential impact flux described by parameters in p {::Proposal} p.tχ ~ instability start time p.τχ ~ e-folding timescale of impact flux p.Fχ ~ initial impact flux

ImpactResetAr( dates::AbsractArray, Vfrxn::AbstractArray, p::NamedTuple, c::NamedTuple Δt::Number,tmax::Number,nᵣ::Integer)

Simulates an impact history from _χ parameters in p, and resets Ar-Ar cooling dates and fractional volumes (Vfraxn) based on impact/crater properties described in c. Impact site morphologies may be described by a conical (cone), parabolic (pbla), or hemispheric (hemi) approximation.

Δt, tmax, and nᵣ respectively define the timestep, model duration, and radial nodes, as in PlntsmlAr function.

Note that Vfrxn is overwritten.

ImpactChron.area_at_depth — Method

area_at_depth(rᵢ::Number, R::Number, x::T) where T<:{Cone,Parabola,Hemisphere}

Calculates the circular area at body radial depth rᵢ (for a body with radius, R) of an impact-affected region, where the volume of the region is approximated by a cone (x::Cone), paraboloid of rotation (x::Parabola), or hemisphere (x::Hemisphere) with a given maximum depth (x.z) and surface radius (x.r).

ImpactChron.csv2dict — Method

csv2dict(filename::String;symbol::Bool=true)

Read a .csv file into a NamedTuple. Assumes columns reflect individual entries, with the first row specifying that entry's key. For NaN-buffered columns (see nt2csv), the NaNs are removed, returning a single-element entry.

A true value for symbol will convert data saved as arrays of Strings into arrays of Symbols.

see also: nt2csv, csv2dict

ImpactChron.csv2nt — Method

csv2nt(filename::String;symbol::Bool=true)

Read a .csv file into a NamedTuple. Assumes columns reflect individual elements, with the first row specifying that element's key. For NaN-buffered columns (see nt2csv), the NaNs are removed, returning a single-element entry.

A true value for symbol will convert data saved as arrays of Strings into arrays of Symbols.

see also: nt2csv, csv2dict

ImpactChron.data2csv — Method

data2csv(filename::String,data)

Save data to a .csv file. data may be a composite type of Dict or NamedTuple or an Array. Accepts fields of any type element. For single-element entries, the rest of the array/table is filled with NaNs.

see also: nt2csv, dict2csv

ImpactChron.dict2csv — Method

dict2csv(filename::String,D::Dict)

Save a Dict to a .csv. Accounts for fields of any type. For single-element entries, the rest of the array/table is filled with NaNs.

see also: nt2csv, data2csv

ImpactChron.histogramify! — Method

histogramify!(dist::AbstractVector, domain::AbstractRange, x::AbstractVector, y::AbstractVector)

In-place histogramify that overwites a pre-allocated vector dist.

see histogramify for details

ImpactChron.histogramify — Method

histogramify(domain::AbstractRange,x::AbstractVector,y::AbstractVector)

Constructs histogram over (linear) midpoints of domain from model outputs in x with corresponding abundances in y. Does not require a constant step in x.

Normalizes the output, such that for output dist ∑ dist[dᵢ] * Δd = 1 (for each dᵢ in the bincenters of domain with step-size Δd), so long as all x ∈ domain. If any x ∉ domain, ∑ dist[dᵢ] * Δd = 1- (∑yₒᵤₜ / ∑yₐₗₗ ) where the corresponding xₒᵤₜ of each yₒᵤₜ is ∉ domain.

Returns only the histogram masses, centers of time bins must be calculated externally. e.g.

Δd = step(domain)
bincenters= LinRange( first(domain)+Δd/2, last(domain)-Δd/2, length(domain)-1)

or see rangemidpoints

ImpactChron.impact_reset_array! — Method

impact_reset_array!(tₓr::AbstractArray, solartime::AbstractRange, impacts::AbstractArray, tcoolₒ::AbstractArray,
                            dates::AbstractArray,Vfrxn::AbstractArray,
                            p::NamedTuple,c::NamedTuple;
                                nᵣ::Integer)

Simulates an impact history from -χ parameters in p (below), and resets Ar-Ar primary planetesimal cooling dates and fractional volumes (Vfraxn) based on impact/crater properties described in c.

Depth-cooling age (relative) abundances are tracked in array tₓr (time x radial depth), with dimensions (time,radius) = (length(solartime),nᵣ), where nᵣ describes the number of radial nodes, as in the PlntsmlAr function.

impacts and tcoolₒ are pre-allocated vectors that respectively record the number of impacts at each time step and the index of the primary cooling date in solartime.

Impact site geometries are codified by types (Cone,Parabola,Hemisphere) defined in c and calculated in the area_at_depth.

Impact flux follows an exponential decay described by parameters in p:

p.tχ ~ instability start time

p.τχ ~ e-folding timescale of impact flux

p.Fχ ~ initial impact flux

ImpactChron.ll_dist — Method

ll_dist(x::AbstractVector,dist::AbstractVector,mu::AbstractVector,sigma::AbstractVector)

Calculate loglikelihood that observations in mu and sigma are drawn from modeled distribution described by x and dist where x contains the bincenters of a normalized histogram dist and mu and sigma respectively contain the mean and 1σ of the observations.

ImpactChron.perturb — Method

perturb(p::NamedTuple,k::Symbol,n::Number)

Return a NamedTuple identical to p, with one field (key k) changed to the value of n. Note that == identity is preserved only if the order of fields in p is as below

Fields: tss,rAlo,R,ta,cAl,Tm,Tc,ρ,Cp,k,tχα,τχα,Fχα,tχβ,τχβ,Fχβ

ImpactChron.rangemidpoints — Method

rangemidpoints(x::AbstractRange)

Calculate a LinRange of the midpoints of each step in x(<:AbstractRange).

ImpactChron.tturbosum — Method

tturbosum(x::AbstractArray)

Fast summing of x with the power of LoopVectorization.jl's @tturbo (multithreaded turbo).