ImpactChron

Documentation for ImpactChron.

ImpactChron.Nrm

ImpactChron.Unf

ImpactChron.lNrm

ImpactChron.ImpactResetAr

ImpactChron.histogramify

ImpactChron.ll_dist

ImpactChron.ll_param

ImpactChron.ll_params

ImpactChron.perturb

ImpactChron.Nrm — Type

Nrm(μ::Float64,σ::Float64)

Immutable struct to describe normally distributed data, reported as mean (μ) and 1σ (σ)

ImpactChron.Unf — Type

Unf(a::Float64,b::Float64)

Immurable struct to describe uniformly distributed data, reported as minimum (a) and maximum (b).

ImpactChron.lNrm — Type

lNrm(μ::Float64,σ::Float64)

Immutable struct to describe lognormally distributed data, reported as log-space mean (μ) and 1σ (σ)

ImpactChron.ImpactResetAr — Method

ImpactResetAr ~ reheat volumes for an exponential impact flux described by parameters in p {::Proposal} p.tχ ~ instability start time p.τχ ~ e-folding timescale of impact flux p.Fχ ~ initial impact flux

ImpactResetAr( dates::AbsractArray, Vfrxn::AbstractArray, p::NamedTuple, c::NamedTuple Δt::Number,tmax::Number,nᵣ::Integer)

Simulates an impact history from _χ parameters in p, and resets Ar-Ar cooling dates and fractional volumes (Vfraxn) based on impact/crater properties described in c. Impact site morphologies may be described by a conical (cone), parabolic (pbla), or hemispheric (hemi) approximation.

Δt, tmax, and nᵣ respectively define the timestep, model duration, and radial nodes, as in PlntsmlAr function.

Note that Vfrxn is overwritten.

ImpactChron.histogramify — Method

histogramify(domain::AbstractRange,x::AbstractVector,y::AbstractVector)

Constructs histogram over (linear) midpoints of domain from model outputs in x with corresponding abundances in y. Requires that each value of x falls within the bounds of domain.

Normalizes the output, such that for output dist ∑ dist[xᵢ] * Δx = 1 (for each xᵢ in the bincenters of x)

Returns only the histogram masses, bincenters must be calculated externally. e.g.

Δd = step(domain)
bincenters= LinRange( first(domain)+Δd/2, last(domain)-Δd/2, length(domain)-1)

ImpactChron.ll_dist — Method

ll_dist(x::AbstractVector,dist::AbstractVector,mu::AbstractVector,sigma::AbstractVector)

Calculate loglikelihood that observations in mu and sigma are drawn from modeled distribution described by x and dist where x contains the bincenters of a normalized histogram dist and mu and sigma respectively contain the mean and 1σ of the observations.

ImpactChron.ll_param — Method

```julia ll_param(x::Number,D::T) -> T ∈ {Nrm,lNrm,Unf}

Calculate the log-likelihood that x is drawn from a distribution D, where D may be... Normal (Nrm), with mean D.μ and 1σ = D.σ

Lognormal (lNrm), with logspace mean D.μ and 1σ = D.σ. Assumes x is already in logspace.

Uniform (Unf) with lowerbound D.a and upperbound D.b (D::Unf always returns loglikelihhood of zero, bounds test is done earlier to speed up code.)

see ImCh_parameters.jl for construction of T-structs

ImpactChron.ll_params — Method

ll_params(p::NamedTuple,d::NamedTuple)

Calculate log-likelihood for a number of proposals in p with corresponding distributions in d

Currently does not evaluate impact (χ) parameters.

ImpactChron.perturb — Method

perturb(p::NamedTuple,k::Symbol,n::Number)

Return a NamedTuple identical to p, with one field (key k) changed to the value of n. Note that == identity is preserved only if the order of fields in p is as below

Fields: tss,rAlo,R,ta,cAl,Tm,Tc,ρ,Cp,k,tχα,τχα,Fχα,tχβ,τχβ,Fχβ