ImpactChron

Documentation for ImpactChron.

ImpactChron.Nrm

ImpactChron.Unf

ImpactChron.lNrm

ImpactChron.csv2dict

ImpactChron.csv2nt

ImpactChron.data2csv

ImpactChron.dict2csv

ImpactChron.downscale!

ImpactChron.histogramify

ImpactChron.histogramify!

ImpactChron.impact_reset_array!

ImpactChron.ll_dist

ImpactChron.ll_param

ImpactChron.ll_params

ImpactChron.nan_regolith!

ImpactChron.perturb

ImpactChron.plntsml_temperature

ImpactChron.radius_at_depth

ImpactChron.rangemidbounds

ImpactChron.rangemidpoints

ImpactChron.tturbosum

ImpactChron.weight_petro_types!

ImpactChron.Nrm — Type

Nrm(μ::Float64,σ::Float64)

Immutable struct to describe normally distributed data, reported as mean (μ) and 1σ (σ)

ImpactChron.Unf — Type

Unf(a::Float64,b::Float64)

Immutable struct to describe uniformly distributed data, reported as minimum (a) and maximum (b).

ImpactChron.lNrm — Type

lNrm(μ::Float64,σ::Float64)

Immutable struct to describe lognormally distributed data, reported as log-space mean (μ) and 1σ (σ)

ImpactChron.csv2dict — Method

csv2dict(filename::String;symbol::Bool=true)

Read a .csv file into a NamedTuple. Assumes columns reflect individual entries, with the first row specifying that entry's key. For NaN-buffered columns (see nt2csv), the NaNs are removed, returning a single-element entry.

A true value for symbol will convert data saved as arrays of Strings into arrays of Symbols.

see also: nt2csv, csv2dict

ImpactChron.csv2nt — Method

csv2nt(filename::String;symbol::Bool=true)

Read a .csv file into a NamedTuple. Assumes columns reflect individual elements, with the first row specifying that element's key. For NaN-buffered columns (see nt2csv), the NaNs are removed, returning a single-element entry.

A true value for symbol will convert data saved as arrays of Strings into arrays of Symbols.

see also: nt2csv, csv2dict

ImpactChron.data2csv — Method

data2csv(filename::String,data)

Save data to a .csv file. data may be a composite type of Dict or NamedTuple or an Array. Accepts fields of any type element. For single-element entries, the rest of the array/table is filled with NaNs.

see also: nt2csv, dict2csv

ImpactChron.dict2csv — Method

dict2csv(filename::String,D::Dict)

Save a Dict to a .csv. Accounts for fields of any type. For single-element entries, the rest of the array/table is filled with NaNs.

see also: nt2csv, data2csv

ImpactChron.downscale! — Method

ImpactChron.downscale!(B::AbstractArray, A::AbstractArray)

Downscales elements of 1-D array A into smaller 1-D array B by summing. Scales the downscaled values in B by the ratio of length(A)÷length(B) to preserve any normalizations.

Requires that length(A) % length(B) = 0.

ImpactChron.histogramify! — Method

histogramify!(dist::AbstractVector, domain::AbstractRange, x::AbstractVector, y::AbstractVector)

In-place histogramify that overwites a pre-allocated vector dist.

see histogramify for details

ImpactChron.histogramify — Method

histogramify(domain::AbstractRange,x::AbstractVector,y::AbstractVector)

Constructs histogram over (linear) midpoints of domain from model outputs in x with corresponding abundances in y. Does not require a constant step in x.

Normalizes the output, such that for output dist ∑ dist[dᵢ] * Δd = 1 (for each dᵢ in the bincenters of domain with step-size Δd), so long as all x ∈ domain. If any x ∉ domain, ∑ dist[dᵢ] * Δd = 1- (∑yₒᵤₜ / ∑yₐₗₗ ) where the corresponding xₒᵤₜ of each yₒᵤₜ is ∉ domain.

Returns only the histogram masses, centers of time bins must be calculated externally. e.g.

Δd = step(domain)
bincenters= LinRange( first(domain)+Δd/2, last(domain)-Δd/2, length(domain)-1)

or see rangemidpoints

ImpactChron.impact_reset_array! — Method

impact_reset_array!(tₓr::AbstractArray, solartime::AbstractRange, impacts::AbstractArray, tcoolₒ::AbstractArray,
                            dates::AbstractArray,Vfrxn::AbstractArray,
                            p::NamedTuple,c::NamedTuple;
                                nᵣ::Integer,Δt::Number)

Simulates an impact history from -χ parameters in p (below), and resets Ar-Ar primary planetesimal cooling dates and fractional volumes (Vfraxn) based on impact/crater properties described in c.

Depth-cooling age (relative) abundances are tracked in array tₓr (time x radial depth), with dimensions (time,radius) = (length(solartime),nᵣ), where nᵣ describes the number of radial nodes, as in the PlntsmlAr function.

impacts and tcoolₒ are pre-allocated vectors that respectively record the number of impacts at each time step and the index of the primary cooling date in solartime.

Impact site geometries are codified by types (Cone,Parabola,Hemisphere) defined in c and calculated in the radius_at_depth.

Impact flux follows an exponential decay described by parameters in p:

p.tχ ~ instability start time

p.τχ ~ e-folding timescale of impact flux

p.Fχ ~ initial impact flux

ImpactChron.ll_dist — Method

ll_dist(x::AbstractVector,dist::AbstractVector,mu::AbstractVector,sigma::AbstractVector)

Calculate loglikelihood that observations in mu and sigma are drawn from modeled distribution described by x and dist where x contains the bincenters of a normalized histogram dist and mu and sigma respectively contain the mean and 1σ of the observations.

ImpactChron.ll_param — Method

```julia ll_param(x::Number,D::T) -> T ∈ {Nrm,lNrm,Unf}

Calculate the log-likelihood that x is drawn from a distribution D, where D may be... Normal (Nrm), with mean D.μ and 1σ = D.σ

Lognormal (lNrm), with logspace mean D.μ and 1σ = D.σ. Assumes x is already in logspace.

Uniform (Unf) with lowerbound D.a and upperbound D.b (D::Unf always returns loglikelihhood of zero, bounds test is done earlier to speed up code.)

see ImCh_parameters.jl for construction of T-structs

ImpactChron.ll_params — Method

ll_params(p::NamedTuple,d::NamedTuple)

Calculate log-likelihood for a number of proposals in p with corresponding distributions in d

Currently does not evaluate impact (χ) parameters.

ImpactChron.nan_regolith! — Method

nan_regolith!(d::AbstractArray,T::AbstractArray,Tmin::Number)

Replace all dates in d with NaN where corresponding peak temperatures in T are less than Tmin Fast with @turbo

ImpactChron.perturb — Method

perturb(p::NamedTuple,k::Symbol,n::Number)

Return a NamedTuple identical to p, with one field (key k) changed to the value of n. Note that == identity is preserved only if the order of fields in p is as below

Fields: tss,rAlo,R,ta,cAl,Tm,Tc,ρ,Cp,k,tχα,τχα,Fχα,tχβ,τχβ,Fχβ

ImpactChron.plntsml_temperature — Method

function plntsml_temperature(time::AbstractArray, radii::AbstractArray; To::Float64, Ao::Float64, λ::Float64, K::Float64, κ::Float64 )

Calculates the evolution of temperature at a range of depths defined for a conductively cooling sphere with thermal conductivity K and thermal diffusivity κ, given ambient temperature To, initial heat production Ao, and heat-producing-element decay constant 'λ'.

Adapted from: Carlslaw & Jäger (1959) and Hevey & Sanders (2006)

ImpactChron.radius_at_depth — Method

radius_at_depth(rᵢ::Number, R::Number, x::T) where T<:{Cone,Parabola,Hemisphere}

Calculates the radius of a circular area at a radial distance of rᵢ from the center of a body with radius R, where the volume of the region is approximated by a cone (x::Cone), paraboloid of rotation (x::Parabola), or hemisphere (x::Hemisphere). x includes a maximum depth (x.z) and surface radius (x.r). When x::Hemisphere, only r is used.

ImpactChron.rangemidbounds — Method

rangebinbounds(x::AbstractRange)

Calculate a LinRange of the linear bounds for each "midpoint" step in x.

ImpactChron.rangemidpoints — Method

rangemidpoints(x::AbstractRange)

Calculate a LinRange of the midpoints of each step in x(<:AbstractRange).

ImpactChron.tturbosum — Method

tturbosum(x::AbstractArray)

Fast summing of x with the power of LoopVectorization.jl's @tturbo (multithreaded turbo).

ImpactChron.weight_petro_types! — Method

weight_petro_types!(v::AbstractArray,T::AbstractArray,d::AbstractArray,petrotypes::NamedTuple)

Reweight volumetric fractions relative to the abundance of each petrologic type in the meteorite record. Takes Vectors of volumetric fraction (v), peak temperature (T), and cooling date (d), as output by PlntsmlAr. Accounts for melted layers (r where isnan(d[i]) is true), by finding NaNs in d, filling these elements of v with zeros. Requires all petrologic types to occupy at least one radial node, otherwise returns all NaNs in in d, which rejects the posterior in MetropolisAr

petrotypes is a NamedTuple of NamedTuples, such that petrotypes = (type#=(T<:Number,p<:Number), ...) where T and p respectively represent maximum temperature and relative abundance in the meteorite record.